Prevented the main room path from breaking up sacrifice pits; ensured they would appear same frequency (see details below).
There was originally a 1/8 chance of sacrifice pits in ALL normal temple levels (Olmec excluded).
However, some of these sacrifice pits would be broken up by the main room path.
Algorithmically calculating the probability of every possible main room path, it was found that 267149/820125 (approximately 1/3) of rooms were compatible with sacrifice pits.
The probability for sacrifice pits was therefore increased by a factor of 3, from 1/8 to 3/8, by checking if the random number was <= 3, instead of == 1.