Fixed #16, removing generation bias for yeti lairs and alien ships; correspondingly adjusted probabilities (see details below).
"They" refers to either the alien ship or the yeti lairs.

25% of time, they had no chance of being compatible to appear.
25% of time, they had 1 chance of being compatible to appear.
25% of time, they had 2 chances of being compatible to appear.
25% of time, they had 3 chances of being compatible to appear.

(0+1+2+3)/4 = 6/4 = 3/2 = Average of 1.5 chances for being compatible to appear in the 4 levels of the ice caves.

The reason why this is considered a bug is the following:

In level 12, they had 0 chances of being compatible to appear.
In level 13, they had 1/4 chances of being compatible to appear.
In level 14, they had 1/2 chances of being compatible to appear.
in level 15, they had 3/4 chances of being compatible to appear.

Their appearance was severely biased towards the last three levels and particularly the 4th level.

Meanwhile now:

100% of the time, they now have 3 chances for being compatible to appear in the 4 levels of the ice caves.

If the probabilities for ship/yeti compatible levels have INCREASED by a factor, the corresponding probability frequencies must also DECREASE in frequency to balance out the change.

Yeti probability has also been untangled from alien ship probability.

Previously, (if compatible), alien ships were 1/10, and if an alien ship had not generated, yeti lairs were IN THEORY 1/6, but actually 9/10*1/6 = 9/60.
If carried over in the same entangled fashion, alien ships would be 1/20, and yeti lairs would be IN THEORY 1/12, but actually 19/20*1/12 = 19/120.

Yeti caves probabilities relative to alien ship probabilities INCREASE as both probabilities (overall probability) DECREASES.

This change makes the ratio between alien ship and yeti lairs probabilities constant as both change proportionally.

Basically, I'm a math nerd, and I want the level generation to be well-ordered, even as someone tweaks with the probability values. I want the probability values to MEAN the same thing at every frequency.